# openapi.yaml
openapi: 3.0.3
info:
  title: Down Dirty 84 AI Tuning MVP API
  version: 0.3.0
servers:
- url: http://localhost:8080
paths:
...
  /api/v1/jobs/{jobId}/analyze:
    post:
      summary: Start analysis for one or more uploaded logs
      description: logUploadIds must reference existing uploads for the same job and authenticated user, and each upload must have kind LOG.
      parameters:
      - in: path
        name: jobId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
      responses:
        '202':
          description: Queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeQueued'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/jobs/{jobId}/validation:
    get:
      summary: Get validation result for a job run
      parameters:
      - in: path
        name: jobId
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: runId
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Validation object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Validation'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/jobs/{jobId}/findings:
    get:
      summary: Get findings for a job run
      parameters:
      - in: path
        name: jobId
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: runId
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Findings object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Findings'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/diffsets/{diffSetId}/export/summary:
    post:
      summary: Export branded summary text for diffset
      parameters:
      - in: path
        name: diffSetId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Summary text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportSummaryResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/diffsets/{diffSetId}/export/csv:
    post:
      summary: Export CSV changelist for diffset
      parameters:
      - in: path
        name: diffSetId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportCsvRequest'
      responses:
        '200':
          description: CSV
          content:
            text/csv:
              schema:
                type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/jobs:
    get:
      summary: List jobs
      responses:
        '200':
          description: OK
    post:
      summary: Create job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Created
  /api/v1/jobs/{jobId}:
    get:
      summary: Get job
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
        '404':
          description: Not found
  /api/v1/jobs/{jobId}/runs/latest:
    get:
      summary: Get latest run status for job
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Latest run status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStatus'
        '404':
          description: Not found
  /api/v1/jobs/{jobId}/runs:
    get:
      summary: List runs for job (newest first)
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Job run list
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/RunStatus'
                required:
                - runs
        '404':
          description: Not found
...
  /api/v1/uploads/jobs/{jobId}:
    get:
      summary: List uploads for job
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
  /api/v1/admin/jobs:
    get:
      summary: Admin list jobs
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
  /api/v1/admin/jobs/{jobId}:
    get:
      summary: Admin get job
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
    patch:
      summary: Admin update job
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    AnalyzeRequest:
      type: object
      description: Analyze payload for a job. logUploadIds must be job-scoped LOG upload IDs.
      required:
      - logUploadIds
      properties:
        logUploadIds:
          type: array
          description: UUIDs of uploads attached to the target job with kind LOG.
          items:
            type: string
            format: uuid
          minItems: 1
        hints:
          type: object
          additionalProperties: true
    AnalyzeQueued:
      type: object
      required:
      - runId
      - status
      properties:
        runId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - QUEUED
    RunStatus:
      type: object
      required:
      - runId
      - status
      - progress
      - errors
      properties:
        runId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - QUEUED
          - RUNNING
          - SUCCEEDED
          - FAILED
          - CANCELED
        progress:
          type: object
          required:
          - pct
          - stage
          properties:
            pct:
              type: number
            stage:
              type: string
        startedAt:
          type:
          - string
          - 'null'
          format: date-time
        finishedAt:
          type:
          - string
          - 'null'
          format: date-time
        errors:
          type: array
          items:
            type: object
            required:
            - code
            - message
            properties:
              code:
                type: string
              message:
                type: string
              details:
                type: object
                additionalProperties: true
    NextLogPlan:
      type: object
      required:
      - planId
      - title
      - routine
      - safetyNotes
      - recommendedLogPreset
      properties:
        planId:
          type: string
        title:
          type: string
        bullets:
          type: array
          items:
            type: object
            required:
            - kind
            - text
            properties:
              kind:
                type: string
              text:
                type: string
        routine:
          type: array
          items:
            type: string
        safetyNotes:
          type: array
          items:
            type: string
        recommendedLogPreset:
          type: string
    Validation:
      type: object
      required:
      - status
      - wbStatus
      - missingChannels
      - recommendedLogPreset
      - nextLogPlan
      properties:
        status:
          type: string
          enum:
          - PASS
          - FAIL
        wbStatus:
          type: string
        missingChannels:
          type: array
          items:
            type: string
        vehicleProfileDetected:
          type: object
          additionalProperties: true
        recommendedLogPreset:
          type: string
        nextLogPlan:
          $ref: '#/components/schemas/NextLogPlan'
    Findings:
      type: object
      required:
      - summary
      - items
      - chartSpecs
      properties:
        summary:
          type: object
          required:
          - blockers
          - warnings
          - info
          properties:
            blockers:
              type: integer
            warnings:
              type: integer
            info:
              type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/FindingItem'
        chartSpecs:
          type: array
          items:
            type: object
            additionalProperties: true
    FindingItem:
      type: object
      required:
      - id
      - severity
      - code
      - title
      - summary
      - evidence
      - actions
      properties:
        id:
          type: string
        severity:
          type: string
          enum:
          - BLOCKER
          - WARN
          - INFO
        code:
          type: string
        title:
          type: string
        summary:
          type: string
        evidence:
          type: object
          required:
          - ranges
          properties:
            ranges:
              type: array
              items:
                type: array
                minItems: 2
                maxItems: 2
                items:
                  type: number
            notes:
              type: string
        stats:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
        charts:
          type: array
          items:
            type: string
        actions:
          type: array
          items:
            type: object
            required:
            - type
            - priority
            - text
            properties:
              type:
                type: string
              priority:
                type: integer
              text:
                type: string
...
    ExportSummaryResponse:
      type: object
      required:
      - text
      properties:
        text:
          type: string
    ExportCsvRequest:
      type: object
      properties:
        includeSuggested:
          type: boolean
          default: false
        minConfidence:
          type: number
          minimum: 0
          maximum: 1
          default: 0.45
...
    CreateJobRequest:
      type: object
      required:
      - serviceType
      - platform
      properties:
        serviceType:
          type: string
        platform:
          type: string
        engineFamily:
          type: string
        vehicle:
          type: string
        ecu:
          type: string
        notes:
          type: string
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: dd84_session